updater.c bug revise
git-svn-id: svn://192.168.1.250/UVC@263 a8a654e3-9019-44d2-b89c-6867c96d4060

+++ RT288x_SDK/source/user/mct_gadget/updater.c
@@ -13,16 +13,16 @@
 #include <stdio.h>
 #include <arpa/inet.h> //inet_ntoa
 #include <sys/types.h> /* open socket */
-#include <unistd.h>	/* open */
+#include <unistd.h>	   /* open */
 #include <sys/stat.h>  /* open */
-#include <fcntl.h>	 /* open */
+#include <fcntl.h>	   /* open */
 #include <sys/ioctl.h> /* ioctl */
-#include <signal.h>	/* SIGPIPE */
+#include <signal.h>	   /* SIGPIPE */
 #include <errno.h>
 #include <pthread.h>
 #include <net/if.h>
 
-#include "nvram.h"
+#include "nvram.h" //catch device infomation
 #include "updater.h"
 #include "CRC.h"
 
@@ -30,43 +30,44 @@
 #define RESET "\033[0m"
 #define FILE_NAME "/var/tmpFW" //   UVC-j5-RX-0.0.0.1.bin
 #define PACKET_LENGTH_MAX 32 * 1024
-#define PORT 55551
+#define TCP_PORT 55551
 #define BROADCAST_IP "255.255.255.255"
 #define BROADCAST_PORT 55550
 
-int tcp_sd, broadcastsd, progress;
+int tcp_sd, broadcastsd, progress, kill_rc, role;
 unsigned int tcptotalget;
 struct sockaddr_in server; //br receform
 struct sockaddr_in form;   //br sendto
+
 void ledcolor(char *color)
 {
-	char led[8];
+	char led[8] = {0};
 	FILE *fp;
 
 	fp = fopen("/tmp/led", "r");
 	if (fp == NULL)
 	{
 		printf("open file error !......\n");
-		pthread_exit(NULL);
+		return;
 	}
 	fread(led, 1, sizeof(led), fp);
 	printf("led = %s\n", led);
 	if (strcmp(color, "red") == 0)
-		if (strcmp(led, color))
+		if (strncmp(led, color, 3))
 		{
 			system("gpio l 52 0 4000 0 1 4000");
 			system("gpio l 14 4000 0 1 0 4000");
 			system("echo 'red' > /tmp/led");
 		}
 	if (strcmp(color, "green") == 0)
-		if (strcmp(led, color))
+		if (strncmp(led, color, 5))
 		{
 			system("gpio l 14 0 4000 0 1 3000");
 			system("gpio l 52 4000 0 1 0 4000");
 			system("echo 'green' > /tmp/led");
 		}
 	if (strcmp(color, "flash") == 0)
-		if (strcmp(led, color))
+		if (strncmp(led, color, 5))
 		{
 			system("gpio l 52 0 4000 0 1 4000");
 			system("gpio l 14 1 1 1 1 4000");
@@ -137,7 +138,7 @@ void printmessage(PJUVCHDR t6_head, PDEVICEINFO t6_info) //print head message
 	{
 		printf("t6_head->Tag = %u\n", t6_head->Tag);
 		printf("t6_head->XactType = %d\n", t6_head->XactType);
-		printf("t6_head->Flags = %u\n", t6_head->Role);
+		printf("t6_head->Role = %u\n", t6_head->Role);
 		printf("t6_head->XactId = %u\n", t6_head->XactId);
 		printf("t6_head->TotalLength = %u\n", t6_head->TotalLength);
 		printf("t6_head->HdrSize = %d\n", t6_head->HdrSize);
@@ -145,9 +146,10 @@ void printmessage(PJUVCHDR t6_head, PDEVICEINFO t6_info) //print head message
 	if (t6_info != NULL)
 	{
 		printf("t6_info->DeviceName = %s\n", t6_info->DeviceName);
-		printf("t6_info->Port = %d\n", t6_info->Port);
 		printf("t6_info->Version = %s\n", t6_info->Version);
 		printf("t6_info->ModelName = %s\n", t6_info->ModelName);
+		printf("t6_info->Port = %d\n", t6_info->Port);
+		printf("t6_info->Manufacture = %d\n", t6_info->Manufacture);
 		printf("t6_info->MacAddress = %02X", t6_info->MacAddress[0]);
 		printf(":%02X", t6_info->MacAddress[1]);
 		printf(":%02X", t6_info->MacAddress[2]);
@@ -186,7 +188,7 @@ void broadcast_create()
 	server.sin_port = htons(BROADCAST_PORT);
 
 	form.sin_family = AF_INET;
-	form.sin_addr.s_addr = htonl(INADDR_BROADCAST); //255.255.255.255
+	form.sin_addr.s_addr = htonl(INADDR_BROADCAST); //255.255.255.255  send BROADCAST setting (INADDR_BROADCAST)
 	form.sin_port = htons(55550);
 	flag = 0;
 	do
@@ -216,14 +218,14 @@ void tcp_create()
 		sleep(1);
 	} while (1);
 
-	printf("test cursor socket tcp_sd = %d ip = %s port = %d \n", tcp_sd, "INADDR_ANY", PORT);
+	printf("test cursor socket tcp_sd = %d ip = %s port = %d \n", tcp_sd, "INADDR_ANY", TCP_PORT);
 	flag = 1;
 	setsockopt(tcp_sd, SOL_SOCKET, SO_REUSEADDR, (const char *)&flag, sizeof(flag));
 	setsockopt(tcp_sd, SOL_SOCKET, SO_RCVTIMEO, &tv, sizeof(tv));
 
 	struct sockaddr_in tcp_addr;
 	tcp_addr.sin_family = AF_INET;
-	tcp_addr.sin_port = htons(PORT);
+	tcp_addr.sin_port = htons(TCP_PORT);
 	tcp_addr.sin_addr.s_addr = INADDR_ANY; //inet_addr(ip[1]);
 
 	do
@@ -257,7 +259,6 @@ int receive_data(int sd, char *modelname, char *version, char *devicename)
 
 	PJUVCHDR t6_head;
 	PDEVICEINFO t6_info;
-
 	FILE *fp;
 
 	//==============================Receive head===================================
@@ -280,7 +281,7 @@ int receive_data(int sd, char *modelname, char *version, char *devicename)
 	t6_head = (PJUVCHDR)buf;
 	printf("rom recvfrom length ret = %d \n", ret);
 	printmessage(t6_head, NULL);
-	if (t6_head->Tag != 1247106627) //JUVC
+	if (t6_head->Tag != JUVC_TAG) //JUVC   1247106627
 	{
 		printf("Tag Error = %d\n", t6_head->Tag);
 		printf("Re-receive Broadcast and check Tag !...........\n");
@@ -291,8 +292,9 @@ int receive_data(int sd, char *modelname, char *version, char *devicename)
 	if (t6_head->XactType == JUVC_TYPE_READY) // 6 ready
 	{
 		t6_head->XactType == JUVC_TYPE_READY;
-		t6_head->Role = JUVC_DEVICE_TYPE_TX;
+		t6_head->Role = role;
 		t6_head->TotalLength = 96;
+		t6_info->Manufacture = 1; //MCT
 		strcpy(t6_info->DeviceName, devicename);
 		strcpy(t6_info->ModelName, modelname);
 		strcpy(t6_info->Version, version);
@@ -312,6 +314,7 @@ int receive_data(int sd, char *modelname, char *version, char *devicename)
 
 	if (t6_head->XactType == JUVC_TYPE_UPDATE_DATA) //need update
 	{
+		system("rm /var/mtd_write.log");
 		tcptotallength = t6_head->TotalLength - 32;
 		if (tcptotallength > (15 * 1024 * 1024) || tcptotallength <= 32) //JUVC
 		{
@@ -358,6 +361,7 @@ int receive_data(int sd, char *modelname, char *version, char *devicename)
 			t6_head->XactType = 7; //XactType = 7 (receive status)
 			t6_head->HdrSize = 32;
 			t6_head->TotalLength = 64;
+			t6_info->Manufacture = 1; //MCT
 			memcpy(buf + 32, &percentage, sizeof(percentage));
 			ret = send(sd, buf, t6_head->TotalLength, 0);
 			printf("send ret = %d\n", ret);
@@ -379,9 +383,10 @@ int receive_data(int sd, char *modelname, char *version, char *devicename)
 }
 int crc_checksum(int sd)
 {
-	int calculate_CRC, receiveCRC, ret;
+	int calculate_crc, receive_crc, ret;
 	char *file_buffer;
 	unsigned char buf[64] = {0};
+	char *ih_name;
 	PDEVICEINFO t6_info;
 	PJUVCHDR t6_head;
 	FILE *fp;
@@ -399,7 +404,7 @@ int crc_checksum(int sd)
 	bzero(file_buffer, tcptotalget);
 	fread(file_buffer, 1, tcptotalget, fp);
 	t6_info = (PDEVICEINFO)(buf + 32);
-	calculate_CRC = crc32(0, (const unsigned char *)file_buffer + sizeof(image_header_t), tcptotalget - sizeof(image_header_t)); //32 file crc
+	calculate_crc = crc32(0, (const unsigned char *)file_buffer + sizeof(image_header_t), tcptotalget - sizeof(image_header_t)); //32 file crc
 	fclose(fp);
 
 	fp = fopen(FILE_NAME, "r");
@@ -409,7 +414,22 @@ int crc_checksum(int sd)
 		return -1;
 	}
 	fread(&head, 1, sizeof(image_header_t), fp); //read data CRC
-	receiveCRC = ntohl(head.ih_dcrc);
+	receive_crc = ntohl(head.ih_dcrc);
+	ih_name = strstr(head.ih_name, "1.");
+
+	printf("ih_magic = %u\n", ntohl(head.ih_magic));
+	printf("ih_hcrc = %u\n", ntohl(head.ih_hcrc));
+	printf("ih_time = %u\n", ntohl(head.ih_time));
+	printf("ih_size = %u\n", ntohl(head.ih_size));
+	printf("ih_load = %u\n", ntohl(head.ih_load));
+	printf("ih_ep = %u\n", ntohl(head.ih_ep));
+	printf("ih_dcrc = %u\n", ntohl(head.ih_dcrc));
+	printf("ih_os = %X\n", head.ih_os);
+	printf("ih_arch = %X\n", head.ih_arch);
+	printf("ih_type = %X\n", head.ih_type);
+	printf("ih_comp = %X\n", head.ih_comp);
+	printf("ih_name = %s\n", head.ih_name);
+	printf("ih_name = %s\n", ih_name);
 	fclose(fp);
 
 	bzero(buf, sizeof(buf));
@@ -419,28 +439,38 @@ int crc_checksum(int sd)
 	t6_head->HdrSize = 32;
 	t6_head->TotalLength = 64;
 
-	char str[20] = {};
-	char str2[15] = {};
-	if (calculate_CRC != receiveCRC)
+	printf("ih_name = %c\n", ih_name[0]);
+	if (ih_name[0] == '1') //MCT
+	{
+		memcpy(buf + 32, "MCT", 4);
+		ret = send(sd, buf, t6_head->TotalLength, 0);
+		printf("send ret = %d\n", ret);
+		printf("ih_name = %c\n", ih_name[0]);
+	}
+	if (ih_name[0] == '2') //J5
+	{
+		memcpy(buf + 32, "j5", 3);
+		ret = send(sd, buf, t6_head->TotalLength, 0);
+		printf("send ret = %d\n", ret);
+		printf("ih_name = %c\n", ih_name[0]);
+	}
+	sleep(3);
+	if (calculate_crc != receive_crc)
 	{
 		printf("Accept File Content is Error !.........\n");
-		printf("calculation CRC32 = %u\n", calculate_CRC);
-		printf("Recive t6_head->DeviceInfo.crc32 = %u\n", receiveCRC);
-		//strcat(str, "CRC = ");
-		//sprintf(str2, "%u", calculate_CRC);
-		//strncat(str, str2, strlen(str2));
+		printf("calculation CRC32 = %u\n", calculate_crc);
+		printf("Recive t6_head->DeviceInfo.crc32 = %u\n", receive_crc);
 		memcpy(buf + 32, "File Content is Error !", 24);
 		ret = send(sd, buf, t6_head->TotalLength, 0);
 		printf("send ret = %d\n", ret);
-		printf("%s\n", str);
 		free(file_buffer);
 		close(sd);
 		return -1;
 	}
 	else
 	{
-		printf("calculation CRC32 = %u\n", calculate_CRC);
-		printf("Recive t6_head->DeviceInfo.crc32 = %u\n", receiveCRC);
+		printf("calculation CRC32 = %u\n", calculate_crc);
+		printf("Recive t6_head->DeviceInfo.crc32 = %u\n", receive_crc);
 		printf("Accept Content is Complete !.........\n");
 		//memcpy(buf + 32, "Content is Complete", 20);
 		memcpy(buf + 32, "Waiting to Burn in", 19);
@@ -451,7 +481,7 @@ int crc_checksum(int sd)
 	free(file_buffer);
 	return 0;
 }
-int write_progress(int sd, int kill_rc)
+int write_progress(int sd, int rc)
 {
 	/*============Write progress 、read Program process and send=================*/
 	int fpeng;
@@ -463,8 +493,8 @@ int write_progress(int sd, int kill_rc)
 	int ret = 0;
 
 	printf("File_TotalLength:%d\n", tcptotalget);
-	printf("kill_rc = %d\n", kill_rc);
-	if (kill_rc != 3)
+	printf("kill_rc = %d\n", rc);
+	if (rc != 3)
 	{
 		snprintf(cmd, sizeof(cmd), "mtd_write -o %d -l %d write %s Kernel &", 0, tcptotalget, FILE_NAME); //FILE_NAME
 		//snprintf(cmd, sizeof(cmd), "/home/ryan/samba/updater/mtd_write -o %d -l %d write %s Kernel &", 0, tcptotalget, "/home/ryan/samba/updater/RusbFW");
@@ -485,7 +515,7 @@ int write_progress(int sd, int kill_rc)
 		sscanf(writebuf, "%d", &progress);
 		printf("progress : %d\n", progress);
 		t6_head = (PJUVCHDR)buf;
-		t6_head->Tag = 1247106627;
+		t6_head->Tag = JUVC_TAG;
 		t6_head->XactType = JUVC_TYPE_UPDATE_STATUS; //XactType = 3 (write status)
 		t6_head->HdrSize = 32;
 		t6_head->TotalLength = 64;
@@ -496,22 +526,20 @@ int write_progress(int sd, int kill_rc)
 		fclose(fp_progress);
 		usleep(500000);
 		if (ret == -1)
-		{
-			return -1;
-		}
+			return -1; //TCP disconnect
 	}
-
 	return 0;
 }
 void *thread_fun(void *data)
-{ // tcp acceptfd program
+{
+	pthread_detach(pthread_self());
 	PJUVCHDR t6_head;
 	PDEVICEINFO t6_info;
 	PTHREADINFO t6_thread;
 	int ret;
 	char buf[128] = {0};
-	char led[6] = {0};
 	FILE *fp;
+
 	t6_thread = *(PTHREADINFO *)data;
 	t6_head = (PJUVCHDR)buf;
 	t6_info = (PDEVICEINFO)(buf + 32);
@@ -532,62 +560,57 @@ void *thread_fun(void *data)
 			ledcolor("red");
 			close_socket(t6_thread->sd, tcp_sd, broadcastsd);
 			printf("receive_data = %d\n", ret);
+			kill_rc = 3;
 			sleep(1);
 			pthread_exit(NULL);
 		}
 		t6_thread->flag++;
 	}
+
 	if (t6_thread->flag == 1)
 	{
 		if ((ret = crc_checksum(t6_thread->sd)) == -1)
 		{
-			t6_head->XactType = 4; //XactType = 4 (CRC)
-			memcpy(buf + 32, "File Content is Error !", 24);
-			send(t6_thread->sd, buf, 56, 0);
-			bzero(led, sizeof(led));
 			ledcolor("red");
 			sleep(1);
+			kill_rc = 3;
 			close_socket(t6_thread->sd, tcp_sd, broadcastsd);
 			printf("crc_checksum = %d\n", ret);
 			pthread_exit(NULL);
 		}
 		t6_thread->flag++;
 	}
+
 	if (t6_thread->flag == 2)
 	{
 		if ((ret = write_progress(t6_thread->sd, t6_thread->kill_rc)) == -1)
 		{
-			t6_head->XactType = 4; //XactType = 4 (CRC)
-			memcpy(buf + 32, "Write_progress Error !", 23);
-			send(t6_thread->sd, buf, 55, 0);
-			bzero(led, sizeof(led));
-			ledcolor("red");
+			//ledcolor("red");
 			printf("write_progress = %d\n", ret);
 			sleep(1);
+			kill_rc = 3;
 			pthread_exit(NULL);
 		}
 	}
+
 	close_socket(t6_thread->sd, tcp_sd, broadcastsd);
 	printf(KGRN "Update Complete!");
 	printf("\n" RESET);
 	ledcolor("green");
-	system("rm /var/mtd_write.log");
 	system("killall -SIGTTIN nvram_daemon");
 	//system("reboot");
-	pthread_detach(pthread_self());
 	pthread_exit(NULL);
 }
 int main(int argc, char *avg[])
 {
-	int recvbytes, sendBytes, fromlen, ret, acceptfd, kill_rc;
+	int recvbytes, sendBytes, fromlen, ret, acceptfd, manuf;
 	unsigned int sin_size, write_length, maxfd;
-	unsigned char buf[128] = {0};
+	char buf[128] = {0};
 	char thread_data[64] = {0};
 	char macaddr[7] = {0};
-	char led[7] = {0};
-	char modelname[32], version[32], devicename[32];
-
-	unsigned char writebuf[4] = {0};
+	char type[3] = {0};
+	char modelname[32], version[32], devicename[32], manufac[32];
+	char writebuf[4] = {0};
 	pthread_t thread_1, thread_2;
 	PJUVCHDR t6_head;
 	PDEVICEINFO t6_info;
@@ -595,8 +618,9 @@ int main(int argc, char *avg[])
 	struct timeval tv;
 	struct sockaddr_in their_addr;
 	FILE *fp, *fp_progress;
+
 	ret = 0;
-	kill_rc = 1;
+	kill_rc = 3; //any
 	acceptfd = 0;
 	tcp_create();
 	broadcast_create();
@@ -604,9 +628,16 @@ int main(int argc, char *avg[])
 	FD_ZERO(&read_sd);
 	signal(SIGPIPE, SIG_IGN);
 	memcpy(macaddr, get_macaddr("rai0"), 6);
-	strcpy(modelname, nvram_get(RT2860_NVRAM, "ModelName"));
+	strcpy(modelname, nvram_get(RT2860_NVRAM, "ModelName")); //catch device ModelName
 	strcpy(version, nvram_get(RT2860_NVRAM, "Version"));
 	strcpy(devicename, nvram_get(RT2860_NVRAM, "HostName"));
+	strcpy(manufac, nvram_get(RT2860_NVRAM, "ManufactureCode"));
+	strcpy(type, nvram_get(RT2860_NVRAM, "Type"));
+	if (strncmp(type, "TX", 1) == 0)
+		role = JUVC_DEVICE_TYPE_TX;
+	else
+		role = JUVC_DEVICE_TYPE_RX;
+	sscanf(manufac, "%d", &manuf);
 
 	while (1)
 	{
@@ -620,22 +651,31 @@ int main(int argc, char *avg[])
 		if (ret == 0) //timeout
 		{
 			printf("Waiting Receive BroadCast !.......  ret = %d\n", ret);
+			fp_progress = fopen("/var/mtd_write.log", "r"); // /var/mtd_write.log  /home/ryan/samba/updater/write_file.txt
+			if (fp_progress == NULL)
+			{
+				printf("File open fail, %d\n", errno);
+				sleep(1);
+				continue;
+			}
+			fread(writebuf, 1, sizeof(writebuf), fp_progress);
+			sscanf(writebuf, "%d", &progress);
+			if (progress == 100)
+			{
+				ledcolor("green");
+				system("killall -SIGTTIN nvram_daemon");
+			}
+			fclose(fp_progress);
+			printf("progress = %d\n", progress);
 		}
+
 		else if (ret == -1)
 		{ //fail
 			printf("Select Error Re-receive Broadcast !...... ret = %d\n", ret);
 			broadcast_create();
-			if (acceptfd > 0)
-			{
-				kill_rc = pthread_kill(thread_1, 0);
-				printf("kill_rc = %d\n", kill_rc);
-			}
-			if (kill_rc != 0) // = 0 alive
-			{
-				tcp_create();
-			}
+			tcp_create();
 			FD_ZERO(&read_sd);
-			kill_rc = 1;
+			kill_rc = 3;
 			ret = 0;
 		}
 		// add tcp connect
@@ -672,14 +712,10 @@ int main(int argc, char *avg[])
 
 			fp = fopen("/var/mtd_write.log", "r"); // /var/mtd_write.log  /home/ryan/samba/updater/write_file.txt
 
-			kill_rc = pthread_kill(thread_1, 0);
-			printf("kill_rc = %d    %d\n", kill_rc, __LINE__);
-
 			if (fp == NULL)
 			{
 				t6_thread->flag = 0;
 				t6_thread->kill_rc = 1;
-				printf("kill_rc = %d    %d\n", kill_rc, __LINE__);
 			}
 			else
 			{
@@ -693,26 +729,27 @@ int main(int argc, char *avg[])
 			printf("kill_rc = %d    %d\n", kill_rc, __LINE__);
 		}
 		// *Broadcast
-		else if (FD_ISSET(broadcastsd, &read_sd))
+		else if (FD_ISSET(broadcastsd, &read_sd) && kill_rc != 0)
 		{
 			fromlen = sizeof(struct sockaddr_in);
 			bzero(buf, sizeof(buf));
 			recvfrom(broadcastsd, buf, 32, 0, (struct sockaddr *)&server, &fromlen);
 			t6_head = (PJUVCHDR)buf;
+			kill_rc = pthread_kill(thread_1, 0);
 
-			if (t6_head->XactType == JUVC_TYPE_DISCOVER) //JUVC_TYPE_DISCOVER =5
+			if (t6_head->XactType == JUVC_TYPE_DISCOVER && kill_rc != 0) //JUVC_TYPE_DISCOVER = 5
 			{
 				system("killall apclient_chkconn.sh");
 				t6_head = (PJUVCHDR)buf;
 				t6_info = (PDEVICEINFO)(buf + 32);
-				t6_head->Tag = 1247106627;
+				t6_head->Tag = JUVC_TAG;
 				t6_head->XactType = JUVC_TYPE_INFO;
-				t6_head->Role = JUVC_DEVICE_TYPE_TX; //TX or RX select
+				t6_head->Role = role; //TX or RX select
 				t6_head->XactId = 204;
 				t6_head->HdrSize = 32;
 				t6_head->TotalLength = 96;
-				t6_info->Port = 55551;
-				t6_info->Manufacture = 1; //MCT
+				t6_info->Port = TCP_PORT;
+				t6_info->Manufacture = manuf; //MCT
 				memcpy(t6_info->MacAddress, macaddr, 6);
 				strncpy(t6_info->ModelName, modelname, strlen(modelname) - 1);
 				strncpy(t6_info->Version, version, strlen(version) - 1);
